requirements.txt
aiogram
aiohttp
tortoise-orm
aiosqlite
